+++ src/soroban/parsers.ts
@@ -23,9 +23,13 @@ export function parseRawEvents(
   return {
     latestLedger: r.latestLedger,
     events: (r.events ?? []).map((evt) => {
+      const clone: Omit<Api.RawEventResponse, 'contractId'> = { ...evt };
+      delete (clone as any).contractId; // `as any` hack because contractId field isn't optional
+
+      // the contractId may be empty so we omit the field in that case
       return {
-        ...evt,
-        contractId: new Contract(evt.contractId),
+        ...clone,
+        ...(evt.contractId !== '' && { contractId: new Contract(evt.contractId) }),
         topic: evt.topic.map((topic) => xdr.ScVal.fromXDR(topic, 'base64')),
         value: xdr.ScVal.fromXDR(evt.value.xdr, 'base64')
       };